+++ ChangeLog.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Unit tests for lb.concurrently and lb.sequentially in test_concurrently.py
 ### Changed
-- Add up to 40 lines of CommandLineWrapper output to debug logs
+- `state` or `settings` traits can be defined directly in a `labbench.Device` class. They are automatically moved into `state` if they include a getter or setter, or otherwise into `settings`.
+- Add first 40 lines of CommandLineWrapper output to debug logs
 - Removed logger warnings when calls to CommandLineWrapper.kill() do not kill any process
 - Tightened the message about a pending exception in lb.concurrently
 - The arguments and return value of sequentially have been corrected to match those in concurrently